+++ CHANGELOG.md
@@ -1,6 +1,17 @@
+## [1.3.2]
+
+add [-v --verbose] flags to `claim` - to print out credential to terminal
+Add Support for `organisation-wallet` version `0.0.10`
+Drop Support for `organisation-wallet` version `<=0.0.9`
+
 ## [1.3.1] (2023.12.18)
 
 Add support to call `claim` and `present` commands with `--url` or `-u` flag to accept url from stdin
+
+## [1.3.0] (2023.12.18)
+
+new format of payload for `/credential` endpoint
+Add error messages on failing to fetch credential-offer or claiming credential
 
 ## [1.2.1] (2023.12.15)